pain: Forget about the loading and saving, the class I tried to make blows up whenever I tried to access its member variable (boolean).
This is probably created from the messes I made for not writing the code properly in the first place, its prolly gonna take to long to fix so I'll just revert things back to how it is supposed to be for now. I am sorry, this whole setup thing is abhorrent I know. Please forgive me...